remove old classes
