+++ app/util.py
@@ -3,10 +3,19 @@ from typing import Dict, Any, Optional, Tuple, Union
 
 import pytz
 from flask import jsonify, Response
+from urlextract import URLExtract
 
 from app.models.base import User
 
 JsonDict = Dict[str, Any]
+
+extractor = URLExtract()
+
+
+def htmlize_note(raw_note) -> str:
+    for url in set(extractor.find_urls(raw_note)):
+        raw_note = raw_note.replace(url, "<a href=\"{url}\" target=\"_blank\">{url}</a>".format(url=url))
+    return raw_note.replace("\n", "<br>")
 
 
 def get_navigation(user: User, current_page: str) -> str:
